+++ tests/grscheller/datastructures/test_ftuple.py
@@ -13,7 +13,8 @@
 # limitations under the License.
 
 from __future__ import annotations
-from grscheller.datastructures.tuples import FTuple as FT, FM
+from grscheller.datastructures.tuples import ftuple as ft, FT
+from grscheller.fp.iterables import FM
 
 class TestFT:
     def test_method_returns_copy(self) -> None:
@@ -24,8 +25,8 @@ class TestFT:
         assert ft1[2] is not None and ft1[2]*2 == ft1[5] == 6
 
     def test_empty(self) -> None:
-        ft1: FT[int] = FT()
-        ft2: FT[int] = FT()
+        ft1: ft[int] = FT()
+        ft2: ft[int] = ft()
         assert ft1 == ft2
         assert ft1 is not ft2
         assert not ft1
@@ -38,7 +39,7 @@ class TestFT:
         assert ft3 is not ft2
         assert not ft3
         assert len(ft3) == 0
-        assert type(ft3) == FT
+        assert type(ft3) == ft
         ft4 = ft3.copy()
         assert ft4 == ft3
         assert ft4 is not ft3
@@ -46,7 +47,7 @@ class TestFT:
         assert ft2[42] is None
 
     def test_indexing(self) -> None:
-        ft0: FT[str] = FT()
+        ft0: ft[str] = FT()
         ft1 = FT("Emily", "Rachel", "Sarah", "Rebekah", "Mary")
         assert ft1[2] == "Sarah"
         assert ft1[0] == "Emily"
@@ -57,8 +58,8 @@ class TestFT:
         assert ft0[0] == None
 
     def test_slicing(self) -> None:
-        ft0: FT[int] = FT()
-        ft1: FT[int]  = FT(*range(0,101,10))
+        ft0: ft[int] = FT()
+        ft1: ft[int]  = FT(*range(0,101,10))
         assert ft1 == FT(0, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100)
         assert ft1[2:7:2] == FT(20, 40, 60)
         assert ft1[8:2:-2] == FT(80, 60, 40)
@@ -69,16 +70,16 @@ class TestFT:
         assert ft0[2:6] == FT()
 
     def test_map(self) -> None:
-        ft0: FT[int] = FT()
-        ft1: FT[int]  = FT(*range(6))
+        ft0: ft[int] = ft()
+        ft1: ft[int]  = ft(range(6))
         assert ft1 == FT(0, 1, 2, 3, 4, 5)
 
         assert ft1.map(lambda t: t*t) == FT(0, 1, 4, 9, 16, 25)
         assert ft0.map(lambda t: t*t) == FT()
 
     def test_foldL(self) -> None:
-        ft0: FT[int] = FT()
-        ft1: FT[int]  = FT(*range(1, 6))
+        ft0: ft[int] = FT()
+        ft1: ft[int]  = FT(*range(1, 6))
         assert ft1 == FT(1, 2, 3, 4, 5)
 
         assert ft1.foldL(lambda s, t: s*t) == 120
@@ -87,8 +88,8 @@ class TestFT:
         assert ft0.foldL(lambda s, t: s*t, start=10) == 10
 
     def test_foldR(self) -> None:
-        ft0: FT[int] = FT()
-        ft1: FT[int]  = FT(*range(1, 4))
+        ft0: ft[int] = FT()
+        ft1: ft[int]  = FT(*range(1, 4))
         assert ft1 == FT(1, 2, 3)
 
         assert ft1.foldR(lambda t, s: s*s - t) == 48
@@ -111,8 +112,8 @@ class TestFT:
             assert False
 
     def test_accummulate(self) -> None:
-        ft0: FT[int] = FT()
-        ft1: FT[int]  = FT(*range(1,6))
+        ft0: ft[int] = ft()
+        ft1: ft[int]  = ft(range(1,6))
         assert ft1 == FT(1, 2, 3, 4, 5)
 
         def add(x: int, y: int) -> int:
@@ -124,12 +125,12 @@ class TestFT:
         assert ft0.accummulate(lambda x, y: x+y, 1) == FT(1)
 
     def test_flatmap(self) -> None:
-        ft0: FT[int] = FT()
+        ft0: ft[int] = FT()
         ft1 = FT(4, 2, 3, 5)
         ft2 = FT(4, 2, 0, 3)
 
-        def ff(n: int) -> FT[int]:
-            return FT(*range(n))
+        def ff(n: int) -> ft[int]:
+            return ft(range(n))
 
         fm = ft1.flatMap(ff)
         mm = ft1.flatMap(ff, type=FM.MERGE)